add test for utilities
